FIX: In plotly_graph_common.py, for _scatterplot force casting to int when calling _interpolate() that cause error with latest versions of numpy.

+++ toulligqc/plotly_graph_common.py
@@ -914,8 +914,8 @@ def _scatterplot(graph_name, dataframe_dict, result_directory, onedsquare=False)
     npoints, sigma = interpolation_points(read_pass_length, 'scatterplot')
     if (len(read_pass_length) + len(read_fail_length)) > npoints:
         pass_ratio = len(read_pass_length) / (len(read_pass_length) + len(read_fail_length))
-        pass_data = _interpolate(read_pass_length, npoints * pass_ratio, y=read_pass_qscore, interp_type="nearest")
-        fail_data = _interpolate(read_fail_length, npoints * (1 - pass_ratio), y=read_fail_qscore,
+        pass_data = _interpolate(read_pass_length, int(npoints * pass_ratio), y=read_pass_qscore, interp_type="nearest")
+        fail_data = _interpolate(read_fail_length, int(npoints * (1 - pass_ratio)), y=read_fail_qscore,
                                  interp_type="nearest")
     else:
         pass_data = [read_pass_length, read_pass_qscore]
